Fix issues #373 and #406 (#432)
* Fix exitMenuItem.text resource for pt_BR

* Refactor saveSettings

This is in preparation of fixing issue #180.

* Refactor loadSettings

This improves error messages. Helps debug issue #180.

* Move readXMLFileFromZipFile to FileOperationsUtil

* Better exception handling for saveSettings

* Revisit exception handling in loadSettings()

* Fix loadSettings for accelerator keys

AcceleratorKeys.setDocument() was unimplemented.

This allows fixing issue #180.

* Change pluskey and numbersign for non-mac

This fixes issue #373

* Translate statusDesktopEntryButton toolTiptext to English and Portuguese

* Catch InaccessibleObjectException in initRegCheckBox()

Registry is inaccessible when opened from Eclipse.

* Add support to accent insensitive search

Also includes major refactoring in StartSearchTask, both to improve
readability and to allow for unit tests.

StartSearchTask could still improve, but this is already too large.

* Undo addition of InaccessibleObjectException to CSettingsDlg.java

Maven project is not finding it. Somehow my setup is compiling
differently than Github's Action

Former-commit-id: 9e3208f2b8236292db0e9bd2347fcac849691be4