fix(askUserAndMaybeSaveChanges): within ZettelkastenView
We need to ensure that settings.getMainDataFile() does not return null before calling toString() on it. One way to do this is by adding a null check before calling toString().

By adding the null check settings.getMainDataFile() != null ?, we ensure that toString() is only called if settings.getMainDataFile() is not null. If it is null, we use the string "null" instead.

This modification will prevent the NullPointerException and ensure that the log message is correctly printed even when settings.getMainDataFile() returns null.

Note: updateSettingsWithCommandLineParams commented out

Former-commit-id: 4dfe4d2dfb582c89a1940e5fd3f6bbe63df2e9c6