test(openEditWindow): debugging showNewEntryWindow RED
`editEntryDlg` is an instance of `EditorFrame` with the title "Neuer Zettel".
This instance represents a frame or dialog for editing a new entry in our application.

From the debug output, we can gather the following information about `editEntryDlg`:

- Class: `de.danielluedecke.zettelkasten.EditorFrame`
- Type: Form for creating a new entry
- Position: (x=312, y=98)
- Size: 831x732
- Layout: BorderLayout
- Title: "Neuer Zettel"
- Resizable: true
- Default close operation: DO_NOTHING_ON_CLOSE
- Root pane: `javax.swing.JRootPane`

This information gives us insights into the state and properties of the `editEntryDlg` object, which can help us diagnose and debug any issues related to its behavior in our application.


Former-commit-id: a0d0de7f502d3a2a7b88086667ad2dffdac7447b