fix(EditorFrame, SelectionListener): ClassCastException
by obtaining the selection model from the JList source in SelectionListener's valueChanged method

In the valueChanged method, e.getSource() returns the source of the event, which is a JList. However, we were trying to cast it to a ListSelectionModel, which causes the ClassCastException.  To fix this issue, we need to use the getSelectionModel() method of the JList to obtain its selection model.