test(ClickListener): log the source of the event. See: #514
To log the source of the event (i.e., the component that triggered the event) in Java, you can use the getSource() method from the ActionEvent class. This method returns the object that fired the event, which you can then log.

In this updated example:

* The ClickListener no longer takes the component name in the constructor. Instead, it determines the source of the event dynamically using e.getSource().

* e.getSource() returns the object that triggered the event, which we check to see if it's an instance of JButton.

* If the source is a JButton, we cast it to a JButton and log its text. This way, the code can handle any number of buttons without hardcoding their names.

Now, whenever a button is clicked, the text of the button that triggered the event is logged. This approach is more flexible and can easily be extended to other types of components.