test(SwingAppWithJavaFX, SwingFXWebView): JFXPanel
"JFXPanel is a component to embed JavaFX content into Swing applications. The content to be displayed is specified with the setScene(javafx.scene.Scene) method that accepts an instance of JavaFX Scene. After the scene is assigned, it gets repainted automatically. All the input and focus events are forwarded to the scene transparently to the developer."

See https://docs.oracle.com/javase/8/javafx/api/javafx/embed/swing/JFXPanel.html